pyproject.toml
[project]
name = "bourbon"
version = "0.1.0"
description = "A general-purpose agent platform starting with exceptional code capabilities"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Bourbon Contributors"}
]
keywords = ["agent", "cli", "coding-assistant", "llm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "anthropic>=0.49.0",
    "openai>=1.66.0",
    "httpx[socks]>=0.28.0",
    "rich>=13.9.0",
    "prompt-toolkit>=3.0.50",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.8.0",
    "toml>=0.10.2",
    "mcp>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
]
stage-b = [
    "pandas>=2.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "jinja2>=3.1.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
]
all = [
    "bourbon[dev,stage-b]",
]

[project.scripts]
bourbon = "bourbon.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/bourbon"
Repository = "https://github.com/yourusername/bourbon"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bourbon"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "integration: marks integration tests",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
